requirements.txt
PySide6>=6.10.2
pypdf>=6.8.0
qtawesome>=1.4.1
pillow>=12.1.1
pymupdf>=1.27.2
pyinstaller>=6.19.0
